fix(module-http): improve hierarchy
* move functionality to lib
