+++ packages/module-http/src/lib/client/index.ts
@@ -1,0 +1,4 @@
+export { HttpClient } from './client';
+export { HttpClientMsal } from './client-msal';
+
+export * from './types';
