+++ packages/module-http/src/lib/index.ts
@@ -1,0 +1,3 @@
+export * from './client';
+export * from './operators';
+export * from './selectors';
