+++ packages/module-http/src/lib/operators/http-response-handler.ts
@@ -1,0 +1,5 @@
+import { ProcessOperators } from './process-operators';
+
+export class HttpResponseHandler<T = Response> extends ProcessOperators<T> {}
+
+export default HttpResponseHandler;
