+++ packages/module-http/src/lib/operators/index.ts
@@ -1,0 +1,5 @@
+export { HttpRequestHandler } from './http-request-handler';
+export { HttpResponseHandler } from './http-response-handler';
+export { ProcessOperators } from './process-operators';
+
+export * from './types';
