+++ packages/module-http/src/lib/selectors/index.ts
@@ -1,0 +1,1 @@
+export { jsonSelector } from './json-selector';
