+++ packages/react-module-http/src/index.ts
@@ -2,6 +2,6 @@ export {
     IHttpClient,
     HttpClientMsal,
     FetchRequestInit,
-} from '@equinor/fusion-framework-module-http';
+} from '@equinor/fusion-framework-module-http/client';
 
 export { useHttpClient } from './useHttpClient';
